Local config file as function param
